+++ src/app/property/[id]/(with-tabs)/detail/InfoBox.tsx
