+++ src/app/property/[id]/(with-tabs)/detail/SectionList.tsx
