+++ src/app/property/[id]/(with-tabs)/detail/components/GrayButton.tsx
