+++ src/app/property/[id]/(with-tabs)/detail/components/ScrollableTabBar.tsx
