+++ src/app/property/[id]/(with-tabs)/detail/section/DealSection.tsx
