+++ src/app/property/[id]/(with-tabs)/detail/section/DescriptionSection.tsx
