+++ src/app/property/[id]/(with-tabs)/detail/section/InfoSection.tsx
