+++ src/app/property/[id]/(with-tabs)/detail/section/LocationSection.tsx
