+++ src/app/property/[id]/(with-tabs)/detail/section/PriceSection.tsx
