+++ src/app/property/[id]/(with-tabs)/review/page.tsx
@@ -1,0 +1,5 @@
+import ReviewTab from "../../review/ReviewTab";
+
+export default function ReviewPage({ params }: { params: { id: string } }) {
+  return <ReviewTab id={params.id} />;
+}
