--- src/app/property/[id]/detail/DetailTab.tsx
@@ -1,17 +1,0 @@
-import InfoBox from "./InfoBox";
-import ScrollableTabBar from "./components/ScrollableTabBar";
-import SectionList from "./SectionList";
-
-interface DetailTabProps {
-  id: string;
-}
-
-export default function PropertyDetailTab({ id }: { id: string }) {
-  return (
-    <div className="flex flex-col gap-2">
-      <InfoBox itemId={Number(id)} />
-      <ScrollableTabBar />
-      <SectionList />
-    </div>
-  );
-}
